moving stoppable thread to utils
